+++ android/vendored/unversioned/@shopify/react-native-skia/cpp/api/JsiSkShader.h
@@ -5,7 +5,7 @@
 
 #include <jsi/jsi.h>
 
-#include <JsiSkHostObjects.h>
+#include "JsiSkHostObjects.h"
 
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdocumentation"
@@ -26,21 +26,7 @@ public:
       : JsiSkWrappingSkPtrHostObject<SkShader>(std::move(context),
                                                std::move(shader)) {}
 
-  // TODO: declare in JsiSkWrappingSkPtrHostObject via extra template parameter?
-  JSI_PROPERTY_GET(__typename__) {
-    return jsi::String::createFromUtf8(runtime, "Shader");
-  }
-
-  JSI_EXPORT_PROPERTY_GETTERS(JSI_EXPORT_PROP_GET(JsiSkShader, __typename__))
-
-  /**
-    Returns the underlying object from a host object of this type
-   */
-  static sk_sp<SkShader> fromValue(jsi::Runtime &runtime,
-                                   const jsi::Value &obj) {
-    return obj.asObject(runtime)
-        .asHostObject<JsiSkShader>(runtime)
-        ->getObject();
-  }
+  EXPORT_JSI_API_TYPENAME(JsiSkShader, "Shader")
+  JSI_EXPORT_FUNCTIONS(JSI_EXPORT_FUNC(JsiSkShader, dispose))
 };
 } // namespace RNSkia