+++ packages/@expo/cli/src/start/server/BundlerDevServer.ts
@@ -17,7 +17,6 @@ import { AsyncNgrok } from './AsyncNgrok';
 import { DevelopmentSession } from './DevelopmentSession';
 import { CreateURLOptions, UrlCreator } from './UrlCreator';
 import { PlatformBundlers } from './platformBundlers';
-import { typescriptTypeGeneration } from './type-generation';
 
 const debug = require('debug')('expo:start:server:devServer') as typeof console.log;
 
@@ -167,12 +166,7 @@ export abstract class BundlerDevServer {
     return false;
   }
 
-  public async startTypeScriptServices(): Promise<void> {
-    return typescriptTypeGeneration({
-      server: this.instance!.server,
-      projectRoot: this.projectRoot,
-    });
-  }
+  public abstract startTypeScriptServices(): Promise<void>;
 
   public async watchEnvironmentVariables(): Promise<void> {
     // noop -- We've only implemented this functionality in Metro.