+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -31,7 +31,7 @@ import {
 } from '../middleware/RuntimeRedirectMiddleware';
 import { ServeStaticMiddleware } from '../middleware/ServeStaticMiddleware';
 import { ServerNext, ServerRequest, ServerResponse } from '../middleware/server.types';
-import { typescriptTypeGeneration } from '../type-generation';
+import { startTypescriptTypeGenerationAsync } from '../type-generation/startTypescriptTypeGeneration';
 import { instantiateMetroAsync } from './instantiateMetro';
 import { getErrorOverlayHtmlAsync } from './metroErrorInterface';
 import { metroWatchTypeScriptFiles } from './metroWatchTypeScriptFiles';
@@ -419,7 +419,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
   }
 
   public async startTypeScriptServices() {
-    typescriptTypeGeneration({
+    startTypescriptTypeGenerationAsync({
       server: this.instance!.server,
       metro: this.metro,
       projectRoot: this.projectRoot,