+++ packages/@expo/cli/src/utils/codesigning.ts
@@ -27,6 +27,8 @@ import { learnMore } from '../utils/link';
 import { env } from './env';
 import { CommandError } from './errors';
 
+const debug = require('debug')('expo:codesigning') as typeof console.log;
+
 export type CodeSigningInfo = {
   keyId: string;
   privateKey: string;
@@ -182,8 +184,8 @@ async function getExpoRootDevelopmentCodeSigningInfoAsync(
   // can't check for scope key validity since scope key is derived on the server from projectId and we may be offline.
   // we rely upon the client certificate check to validate the scope key
   if (!easProjectId) {
-    Log.warn(
-      `Expo Application Services (EAS) is not configured for your project. Configuring EAS enables a more secure development experience amongst many other benefits. ${learnMore(
+    debug(
+      `WARN: Expo Application Services (EAS) is not configured for your project. Configuring EAS enables a more secure development experience amongst many other benefits. ${learnMore(
         'https://docs.expo.dev/eas/'
       )}`
     );