+++ tools/src/vendoring/legacy.ts
@@ -590,6 +590,24 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
         sourceAndroidPackage: 'com.reactnativecommunity.webview',
         targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.webview',
       },
+      {
+        sourceAndroidPath: 'android/src/oldarch/com/reactnativecommunity/webview',
+        cleanupTargetPath: false,
+        targetAndroidPath: 'modules/api/components/webview',
+        sourceAndroidPackage: 'com.reactnativecommunity.webview',
+        targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.webview',
+        onDidVendorAndroidFile: async (file: string) => {
+          const fileName = path.basename(file);
+          if (fileName === 'RNCWebViewPackage.java') {
+            let content = await fs.readFile(file, 'utf8');
+            content = content.replace(
+              /^(package .+)$/gm,
+              '$1\nimport host.exp.expoview.BuildConfig;'
+            );
+            await fs.writeFile(file, content, 'utf8');
+          }
+        },
+      },
     ],
   },
   'react-native-safe-area-context': {
@@ -646,6 +664,13 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
         targetAndroidPath: 'modules/api/components/datetimepicker',
         sourceAndroidPackage: 'com.reactcommunity.rndatetimepicker',
         targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.datetimepicker',
+      },
+      {
+        sourceAndroidPath: 'android/src/paper/java/com/reactcommunity/rndatetimepicker',
+        targetAndroidPath: 'modules/api/components/datetimepicker',
+        sourceAndroidPackage: 'com.reactcommunity.rndatetimepicker',
+        targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.datetimepicker',
+        cleanupTargetPath: false,
       },
     ],
     warnings: [